Alter code for 1. Compose for seperate Comose of Kafka and Spark cluster 2. Dockerfiles to take ZIPped /Tar file of hadoop and kafka code/binaries 3. Sub-pub and kafka consumers code working and tested.

+++ hadoop/remhostcp.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 
-rm remote_hosts.txt
+if [ -f "remote_hosts.txt" ] ; then
+ echo " Previous remote_hosts.txt exists. Deleting it ..."
+ rm remote_hosts.txt
+fi 
 
 # List of host names
 hosts=("hadoop-nameNode-1" "hadoop-dataNode-1" "hadoop-dataNode-2" "hadoop-dataNode-3")
